final_app/requirements.txt

flask>=3.0
pillow         
numpy
torch         
ultralytics       
python-dotenv   
psycopg2-binary 
redis         
